Refactor deck check route by extracting validation logic
- Extract deck validation logic to `lib/deckValidation.ts` utility file
- Create `useDeckValidation` hook for validation state management
- Create `useDeckList` hook for deck list operations and state
- Reduce route component size by ~100+ lines while maintaining functionality
- Improve code organization and reusability for future features
- Maintain full TypeScript support and existing behavior

This refactoring improves maintainability and makes the validation logic
reusable across the application.
